chore: move constants to src
