--- src/app/utils/constants.js
@@ -1,1 +1,0 @@
-export const DomainApi = "/api/dummy/learningpaths";
