+++ src/constants/index.js
@@ -4,3 +4,4 @@ export const tags = {
   AppDevelopment: ["Android", "IOS", "Swift", "Kotlin", "Java", "React Native"],
   "AI/ML": ["Tensorflow", "Deep learning", "NLP"],
 };
+export const DomainApi = "/api/dummy/learningpaths";
